refactor(balanco): mover Balanco Manual pra aba Seminovos de /admin/estoque
Nicolas: 'mas isso deveria ficar na aba de seminovos'.

Antes: secao 'Balanco Manual' ficava em /admin/usados, que o Nicolas nao
encontrava. A tela /admin/usados eh sobre avaliacoes de seminovos no
formulario de troca, nao sobre itens em estoque.

Agora: secao foi movida pra /admin/estoque na aba Seminovos, logo acima
da listagem de produtos. Muito mais intuitivo — eh onde voce ve os
seminovos em estoque e aplica o balanco.

Extraido o componente pra components/admin/BalancoSeminovosSection.tsx
(reutilizavel, suporta modo colapsado ou sempreAberta).

Comportamento:
- Secao colapsada por padrao (clica pra abrir o ▼)
- Mostra modelos agrupados por modelo+armazenamento
- Checkbox por grupo + 'Selecionar todos'
- Botao 'Fazer balanco dos selecionados (N)'
- Modal de confirmacao com preview antes de aplicar
- Nao interfere com o botao 'Balancear Precos' ja existente (eh item-a-item)
